% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_size_procfreq.R
\name{local_size_sas_freq}
\alias{local_size_sas_freq}
\title{Local size of the SAS Proc FREQ exact test}
\usage{
local_size_sas_freq(nuisance, .odds_ratio, .m, .n, .alpha)
}
\arguments{
\item{nuisance}{The nuisance parameter \eqn{p_0}: success probability in
group 1 under \eqn{H_0}. Must be in \eqn{[0, 1]}.}

\item{.odds_ratio}{The null hypothesis odds ratio \eqn{\theta_0}. No default.}

\item{.m}{Number of trials in group 1.}

\item{.n}{Number of trials in group 2.}

\item{.alpha}{Nominal significance level \eqn{\alpha}. No default.}
}
\description{
Computes the unconditional rejection probability (local size) of the SAS
Proc FREQ exact test for \eqn{H_0}: OR = \eqn{\theta_0} at nuisance
parameter \eqn{p_0}, where
\eqn{p_1 = p_0 / (p_0 + \theta_0 (1 - p_0))}. For each possible table
\eqn{(i, j)}, rejection is determined by \code{sas_procfreq_pvalue()}: the
sum of Fisher non-central hypergeometric probabilities (given
\eqn{T = i + j} and OR = \eqn{\theta_0}) that are no greater than the
observed probability, compared against \eqn{\alpha}.
}
\keyword{find}
\keyword{hypothesis}
\keyword{local}
\keyword{procfreq}
\keyword{sas}
\keyword{size}
\keyword{test}
