% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_size_procfreq.R
\name{sas_freq_pvalue}
\alias{sas_freq_pvalue}
\title{Local size of the SAS Proc FREQ "exact" test
The SAS Proc FREQ Fisher exact p-value sums all hypergeometric probabilities
(under H0: OR = 1, i.e. the central hypergeometric) that are no greater than
the observed table probability. Reject H0 if that sum <= alpha. This uses
only base-R dhyper(),  no extra packages needed. Because the null
distribution is always the central hypergeometric (OR = 1), no test frame is
required.}
\usage{
sas_freq_pvalue(s, t, m, n)
}
\arguments{
\item{nuisance}{Nuisance parameter.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}
}
\description{
Local size of the SAS Proc FREQ "exact" test
The SAS Proc FREQ Fisher exact p-value sums all hypergeometric probabilities
(under H0: OR = 1, i.e. the central hypergeometric) that are no greater than
the observed table probability. Reject H0 if that sum <= alpha. This uses
only base-R dhyper(),  no extra packages needed. Because the null
distribution is always the central hypergeometric (OR = 1), no test frame is
required.
}
\keyword{find}
\keyword{hypothesis}
\keyword{local}
\keyword{procfreq}
\keyword{sas}
\keyword{size}
\keyword{test}
