% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_procfreq_pvalue.R
\name{sas_procfreq_pvalue}
\alias{sas_procfreq_pvalue}
\title{P-value from the SAS Proc FREQ exact test used in the local size calculation
The SAS Proc FREQ Fisher exact p-value sums all hypergeometric probabilities
(under H0: OR = 1, i.e. the central hypergeometric) that are no greater than
the observed table probability. Reject H0 if that sum <= alpha. This uses
only base-R dhyper(),  no extra packages needed. Because the null
distribution is always the central hypergeometric (OR = 1), no test frame is
required.}
\usage{
sas_procfreq_pvalue(s, t, m, n)
}
\arguments{
\item{s}{Integer input. Number of successes in m. No default.}

\item{t}{Integer input. Number of successes in n. No default.}

\item{m}{Integer input. Total number of trials m. No default.}

\item{n}{Integer input. Total number of trials n. No default.}
}
\description{
P-value from the SAS Proc FREQ exact test used in the local size calculation
The SAS Proc FREQ Fisher exact p-value sums all hypergeometric probabilities
(under H0: OR = 1, i.e. the central hypergeometric) that are no greater than
the observed table probability. Reject H0 if that sum <= alpha. This uses
only base-R dhyper(),  no extra packages needed. Because the null
distribution is always the central hypergeometric (OR = 1), no test frame is
required.
}
\keyword{find}
\keyword{hypothesis}
\keyword{local}
\keyword{procfreq}
\keyword{sas}
\keyword{size}
\keyword{test}
